shoppingmall_back/requirements.txt
beautifulsoup4==4.8.1
cryptography==2.8
Django==3.0
lxml==4.4.2
PyJWT==1.7.1
requests==2.22.0
selenium==3.141.0